+++ src/structures/vroom/raw_route.h
@@ -70,6 +70,8 @@ public:
 
   bool has_pickup_up_to_rank(const Index rank) const;
 
+  const Amount& max_load() const;
+
   // Check validity for addition of a given load in current route at
   // rank.
   bool is_valid_addition_for_capacity(const Input&,